WIP - Implement HTTP-server based key agent
Create a key agent which runs as a daemon process.
The key agent operates similarly to ssh-agent. New instances
of the 1pass CLI connect to it and can add keys to it.

When an agent is already running, new CLI instances can send
the agent requests to decrypt (and in future encrypt) item data.