Item list layout improvements
Use an overflowing <div> rather than just stretching the body
for the item list.

This fixes an issue where the search bar would jump down from
the top of the screen when focused on iOS due to the way iOS
handles position:fixed.

This also fixes an issue where the details view was shown
off the top of the screen when switching to it whilst scrolled
in the item list.