HTML LocalStorage-based VFS
Add a VFS implementation based on the LocalStorage API.

This implementation stores file contents and directory indexes
as keys prefixed with a given root key. Each key stores either a directory
index as a JSON string or the contents of a file.

Directory indexes are an array of file metadata entries. Subdirectories
contents by either be included in the directory index for the parent
directory or may be stored in a separate key.

eg:

[
 { name: 'filea', key: '<filea key>' },
 { name: 'dira' , entries : [ { name : 'fileb', key: '<fileb key>' } ] },
 { name: 'dirb' , key: '<dirb key>' }
]

The directory index for the root dir is stored in the root key passed
to the localstorage.FS constructor.

The LocalStorage VFS is tested using the existing VFS tests which
have been modified to be usable with different VFS implementations.