Add siteinfo/ module for fetching hi-res icons for websites/
siteinfo/ contains code to fetch icons for URLs from:
 - Common icon paths (favicon.ico, apple-touch-icon)
 - App, social media, schema.org and other resources linked from
   a site's main page

Images in ICO, PNG, JPEG and BMP format are supported.

 - ico.ts implements a .ico format handler which can parse .ico files
   containing data in the most common bmp format and extract the individual
   icons in .bmp format.

 - image.ts implements image size metadata extraction for BMP, JPEG and
   PNG files

 - service.ts implements a service which will query the standard icon
   paths ('favicon.ico', 'apple-touch-icon.png') for a site and also
   fetch the HTML content and extract site links from <link> and
   <meta> tags.

Images from external domains are fetched via a UrlFetcher interface,
which can either be a front-end to a web service in the case of a
web-only front-end where cross-origin requests are not allowed,
or priviledged APIs in the case of a browser extension or native app.