Unify the three CLI 'show' commands and only show one item
Replace the separate 'show-overview', 'show-json' and 'show' commands
with a single command which takes a 'format' argument.

Also change the 'show' command to only show the content of
one item and prompt to select if there are multiple matches
in the same way as 'edit' and 'copy'.

This is more consistent with those commands and seems more useful
in practice.