Show items matching the current tab if the search filter is empty
When the search filter is empty, default to showing only items
matching the URL of the current tab.

If no items appear to match the current tab then the full list is
shown for the user to search or browse.

If the user starts typing in the search box then the full list of
items is searched for matches.