Initial implementation of form autofill in Firefox addon
Add a button to the item details view page to trigger
autofill. When clicked, the <input> fields on the current page
are collected, the web front-end matches them against the item
and sends back a list of auto-fill entries to the page via
the main add-on script.

This version works for a few sites (Evernote, Dropbox)
but the username and password fields are not detected correctly
for many others.