Switch from item detail view back to main item list when changing pages
When changing pages, the previously selected item is likely
not to be relevant any longer, so reset the selection.

It would be better here not to reset the selection
if the best URL match remained unchanged after the URL change,
eg. when following a link to another on the same domain.