Fix autofill no longer working after activating link in tab
When navigating to a new page in a tab which has been autofilled,
the existing worker is detached and is no longer active.

Remove the worker from the tab ID => worker mapping in this
case so a new worker is set up when the user next activates
autofill in the tab.