Convert messages between content and add-on scripts to find form fields to use the RPC implementation
This works around an issue where repeated calls to self.port.once() can
fail due sporadically due to an issue in the way 'for each ()' handles
deletion from the collection during iteration - see https://github.com/mozilla/addon-sdk/pull/1557
for details.

It also ensures that requests and responses to find fields in the
current page match up. Previously this was not guaranteed.

This currently copies the `rpc.ts` file into addons/firefox/src/ as part of the build
so that main.ts can reference it via require('./rpc'), otherwise the
typescript compiler will recreate as much of the directory hierarchy as
necessary in the output lib/ folder.