Avoid window RPC ports receiving messages sent by themselves
When two ports are created for the same window, as is the case
in the Firefox add-on where one port is in the app frontend
and another is the extension's content script, the ports
on both sides were receiving all of the messages they were sending.

Fix this by introducing a 'send tag' and 'receive tag' to
WindowMessagePort. Ports will include the send-tag with all outgoing
messages and only respond to incoming messages where the tag matches
the port's receive-tag.