Add 'repair' command to CLI to repair item details
The 'repair' command loads each item in a vault, runs checks on it
and prompts to repair the item if any issues are found.

The initial use for this is a problem I encountered where the 'Location'
field for many of my items was wiped, although the URLs were still
present in the encrypted data. This issue occurred around the time I
used 1Password v4 for Android, although I have yet to verify that was to
blame.