Add Firefox add-on SDK to build dir
Add Firefox add-on SDK v1.16 submodule.
This provides the cfx tool that will be used to create
and test the Firefox add-on as part of the Travis CI build.

In future we'll be able to use cfx' successor jpm tool
installed via npm. That currently appears to require
Firefox >= 33 however so we'll stick with cfx to support
the current stable version of Firefox