Restructure app UI to support Chrome extension's background/popup page architecture
In Chrome, popups for browser action pages are unloaded whenever
the popup is dismissed.

With the previous structure where the app was initialized and all state
maintained in the popup HTML page, this meant that the passcards app was
being reloaded every time the browser action was used.

This commit restructures the app so that the bulk of the app
lives in a background page which can live whilst the popup
is not active.

When the popup is shown, the app's main component is rendered into
the DOM element in the popup.

 * Separate initialization of the app from rendering of the app's main
   view. In the Chrome extension the app is rendered into a different
   document than the background page where the app lives.

 * Remove several dependencies on the global 'document' object
   in components in favor of using the document that the component
   is currently mounted into.

 * Set up Dropbox OAuth receiver script for Chrome extension to handle
   Dropbox login.

 * Use key codes rather than KeyboardEvent.key string identifiers for
   compatibility with Chrome and Firefox