Fix page content script
The recently added env.ts dependency added to page.ts caused
browserify to add a dependency on Node's process module however
due to the --no-builtins flag being passed to browserify, the
process stub was not added.

Since the process object stub is small, we can afford to include
it in the page content script and resolve the issue by building
the script bundle with browserify builtins enabled.