Reduce strictness of item matching when filtering items
Instead of requiring the pattern to occur as a complete
phrase within one of the main item metadata fields,
split the pattern into whitespace-separated clauses
and require each clause to match one of the metadata
fields.

The shell command parsing is re-used to parse the
search pattern to stricter matching can be achieved
using quotes.

eg. 'acme goog' will now match an item titled
 'ACME Ltd. - Google Account' but '"acme goog"'
would not.