Base64-encode message (plain|cipher)text in messages with the key agent
The messages sent to the key agent were JSON but the
plaintext/ciphertext were binary strings. This happened to work in the past
because they were being UTF-8 encoded when sent and UTF-8 decoded when
received. Following the fix in streamutil.readAll() to decode as binary
this broke.

Use base64 encoding as the standard way to exchange binary data in JSON
messages.