Add a Material Design-style ripple effect to list items and buttons
The ripple is currently implemented as a <canvas> element which
begins an animation on mouse-down.

The ripple component currently doesn't directly respond to touch
events, but in the context of the passcards app this is not a problem
due to use of the Fastclick script from the FT labs which causes
touch start events to be handled the same as mouse-down events.