Decouple onepass.Item from onepass.Vault
In preparation for introducing a local cache of items,
decouple onepass.Item from the vault.

Instead of being associated with a Vault, items are now
associated only with an item_store.Store interface which
may either be the vault in Dropbox or a local store
(eg. in an IndexedDB).

Also fix several type errors in the Agile Keychain interfaces
and vault tests.