Add partial support for detecting conflicting writes in VFS.write()
Support throwing an error from VFS.write() when writing an update
to an existing file based on a non-current version of the file.

This is work towards fixing a problem in onepass.Vault.saveItem()
where concurrent updates to the contents.js file can clobber each
other and result in items being lost from the index.

Add a 'revision' property to the result of VFS.stat() and
a 'parentRevision' property to a new options struct for VFS.write().
If the 'parentRevision' property is set in VFS.write() and does
not match the current revision of the file, the update fails
with a vfs.ConflictError.

In Dropbox, this maps to the API's mechanism for detecting conflicting
updates.

This system currently isn't completely reliable for the Node FS-based
VFS since the file mtime + size is used as a revision and the mtime only
has a resolution of 1 second in Node 0.10 and on OS X across all Node
versions.