Fix clobbering of contents.js updates by concurrent Item.save() calls
When multiple items are being saved in an Agile Keychain vault,
the updates to contents.js could clobber each other.

Fix this by only allowing one update of contents.js by a onepass.Vault
instance at a time.

When Item.save() is called, the update is enqueued and process to
write an updated contents.js file is scheduled. If Item.save() call
happens concurrently, it will either get picked up by the
already-scheduled update or another one which will happen when the
first completes.

When the keychain is stored in a remote VFS, eg. Dropbox, this also
reduces the number of API calls made to update the contents.js file.