Fix sporadic failure in VFS write conflict test
When performing the conflicting writes, if the write of the
string 'conflict-v2' happened first then the other write attempt
would not see a conflict because the (second-resolution mtime, file
size) tuple which make up the revision did not change.

Work around the problem by using different file contents of different
lengths for the initial write and the two conflicting writes.