Make the IndexedDB implementation slightly more efficient by re-using transactions
Group all operations on the same object store which happen in the same
turn of the event loop into a single transaction.

IDB doesn't provide a way to close transactions explicitly, aside from
aborting them, so there isn't a mechanism to group requests spanning
multiple turns of the event loop into a single transaction
as far as I know.