Implement syncing of items from store to vault using revision history
Use item revision history maintained by the local store as
the basis for syncing.

When a change is detected in either local or cloud stores, fetch
the last-synced revision of an item and use it to decide whether
to use the local or cloud versions of an item or do a 3-way merge.