Implement ordered set and list diff-ing functions
Implement ordered set and list diff-ing functions
for use with 3-way merging of item section, URL and field lists.

The implementation uses the 'adiff' package to perform
the initial list diffing but we implement the patching
and merging ourselves in order to:

 1) Add support for diffing/patching of ordered sets,
    represented as arrays of unique elements

 2) Have a diff implementation with reasonable internal documentation
    for debugging.

I looked at a number of diffing libraries before implementing
this, including adiff, xdiff, jsondiffpatch and diff-merge-patch.

 - xdiff is the simplest package but lacks control over merge resolution
   in the event of conflicts.
 - jsondiffpatch lacks 3-way merging support
 - diff-merge-patch has all the features needed but the README warns
   of bugs and there is at least one basic one documented at
   https://github.com/mirkokiefer/diff-merge-patch/issues/2

   The implementation is non-trivial to read/understand/fix.