Implement merging of URL fields with the same labels
The 'label' field of URL entries is not unique and
the official 1Password app for iOS at least defaults
to using 'website' as the label for each new URL entry.

Use the combination of label+url as the field key.
This allows us to detect moves of URL fields in the list
during merges. Any changes to existing fields are handled
as removal + insert rather than modify (or move+modify)