Load react-style CSS styles in the Chrome extension
react-style injects the compiled CSS into a <style>
element in the document.

In the Chrome extension the app is rendered into a different
document (belonging to the popup view)
than the one which the app code runs in (belonging to the background
page), so we need to copy the <style> element from the background
page document to the popup view's document.