Fix error focusing master password field in Chrome extension
An error occurred if the UnlockView component was mounted whilst
unlocked. This can happen in the Chrome extension where the UI
is mounted when the popup view is shown but the state of the app,
including whether it is unlocked or not, is persisted when the popup
view is dismissed.