Fix tslint errors
One non-obvious change here is that ItemList.setFocus() is used
by a mixin but tslint cannot see that usage, so the method
needs to be marked public to avoid a warning about an unused method.

Marking the method as protected would probably be a better way
to mark internal methods for use by mixins.