New-look details view with 'hero' transition on entrance/exit
Replace the slide animation when transitioning to/from the details
view with a hero transition where details view expands from
the selected item with the icon moving up to a header at the
top of the view and the item details fading in below the header.

This also includes optimizations for item list re-rendering
by avoiding re-filtering the item list when a state change
doesn't affect the list of matching items.