Improve smoothness of rounding for icon images in item list
Previously the icon images were being clipped twice, once
by border-radius applied to the image itself and again
by border-radius applied to the container.

This resulted in aliasing along the edges of icons which
filled the 48x48px container on non-high DPI displays.

Applying only border-radius to the outer container
for icons which fill the container gives a smooth
anti-aliased outline.