Work around an issue with border-radius clipping in Blink/WebKit
On iOS 8, the circular clipping of site icon images disappeared
during the ripple animation in the item list and in Chrome
the clipping disappeared during the animated transition of
the edit/save toolbar icon when switching between view and
details mode in the list.

This looks like it may be related to
https://code.google.com/p/chromium/issues/detail?id=430184

Forcing the containing <div> for the icon's image into its
own compositing layer resolves the problem.

Issue is not reproducible in current Firefox nightly.