Adapt menu display to screen
For small screens (<400 device independent pixels),
the menu occupies a full edge of the screen as menus
do in Google's recent iOS apps on iPhones.

For larger screens, align the menu's top left corner
with the top-left corner of the element that launched
the menu if it will fit, or pick one of the other
corners of the element/menu to align otherwise.