Improve username/account autofill compatibility
Many username/account input fields do not contain
a placeholder, name or ID that matches one of the keys
that the autofill logic's heuristics look for.

This adds a fallback mechanism for identifying the login field.
If a form on the page contains a password field but no corresponding
account field is found then the first non-hidden text field
in the same form is used.

Input fields which are not contained within any form element,
as on the iCloud.com login screen for example,
are put into a single fake 'root' form.

This fixes account autofill on iCloud.com, GitHub.com and many other sites.