Preserve selected item when closing and re-opening Chrome extension popup
When the app was re-rendered, the AppView was created
with state.isLocked initially set to true and then immediately
updated with state.isLocked set to false, resulting in the selected
item being reset.

Fix this by creating the AppView with the saved initial state
rather than with the default state and then immediately
updating it to the saved state.