Move all state related to open item store out of AppView component
In the Chrome extension, the popup page is destroyed when the popup
is closed and the AppView component gets unmounted.

In order to restore the state of the app in a clean way when re-opening
the popup, move all state related to the opened store - eg. cache of
decrypted item overviews, selected items to a separate store.

This fixes several bugs related to incorrect restore of app state
when re-opening the Chrome popup view.

This is also a path to using a more Flux-like way of managing
the app's state.