Make VFS.mkpath() fail if the target directory already exists
Previously the localstorage and Node implementations of VFS.mkpath()
had different behaviour to the Dropbox implementation if the path
already existed. The Node implementation would succeed, like 'mkdir -p'
and the Dropbox implementation would fail.

Make the Node/localstorage behaviors consistent with the Dropbox one.
The Dropbox behavior is more useful because it can be used to make
agile_keychain.Vault.createVault() fail if the vault already exists.