Adapt other tests to change in VFS.mkpath() behavior
mkpath() now fails if the target directory already exists.
Modify several tests to account for this.

Also make FileVFS consistent about how it handles paths.
Previously arguments to FileVFS methods could be either
relative to the root or absolute paths. For consistency
with other VFS implementations, treat all paths as relative
to the root passed to the constructor.