Fix agent_server launch failure under iojs
The key agent server first spawns a new process, waits for
it to write its PID to a file and then reads the PID from that
file to confirm the launch.

Previously there was an assumption that the PID was readable
as soon as the first watch event notification was received for the PID
file. Testing under iojs 1.5 this is not the case.

Keep watching the PID file until a PID is readable from it.