Fix bug where item icon did not update correctly if an icon was immediately available
BasicIconProvider never emitted an update if it had no
icon in its local cache for a given item URL but the
icon data provider was able to supply a result immediately.

This resulted in the details view for an item with no
icon displaying the 'loading' placeholder icon instead
of the 'no icon available' placeholder icon.

Unify the two code paths which handled the cases where
an icon is immediately available from the data provider and
the case where it becomes available after a delay (eg. after a query
to a remote API).