Add utility script to generate an 'umbrella' reference file
Add script to generate an umbrella reference file that includes
all of the TypeScript definitions from a directory containing
type definition modules installed via tsd.

The 'tsd' utility can generate a similar file itself but the current
version is buggy when updating an existing input file, mangling
paths in existing <reference> lines.

This utility script replaces the whole file if it already exists
and sorts the reference lines in the output.