Add method to retrieve only the states of all items in the store
Unlike listItems(), listItemStates() only provides
the ID, last update time and whether the item is a tombstone.

For vaults in the Agile Keychain format, the ID
and last update time can be obtained from the
metadata for .1password files and only the list
of tombstones is needed from the contents.js index
file.

This is a step towards fixing problems arising when
contents.js and the associated .1password file are out
of sync by using only the .1password file as
the source for item metadata.

In order to do this whilst enabling efficient incremental
sync, the list of non-deleted items in the store and their
last update times will be taken from the metadata for
the .1password files instead of contents.js. The list of
deleted items could initially still come from contents.js
but in future we could use the deletion history that Dropbox
provides via its /delta API.