Bump schema version and wipe exiting in-browser DB cache
The schema for the IndexedDB cache changed by storing
separate item revisions for local and remote stores.

The local revision is generated based on item contents
and the parent revision in a similar manner to Git.

The remote revision is taken from the revision metadata
property from Dropbox for the corresponding .1password
file.