Avoid failing the whole sync if an individual item fails
If syncing a particular item fails, eg. due to a service
issue with the cloud provider's API or an item being concurrently
deleted in the cloud then just log a failure for that item,
since the remaining items can be synced.

If there is a condition which causes every item to fail,
such as the encryption keys having changed then this
will log a sync in which all items failed.