Generate input change events when autofilling input fields
For login forms which do not enable the login until the input
fields are validated, 'input'/'change' events need to be emitted
from the input fields after they are autofilled for the login
forms to update. Just setting the HTMLInputElement.value property
does not result in the relevant event handlers being called.

As a result, it was not possible to login to pages such as iCloud.com
after the login inputs were autofilled until the user manually pressed
a key in the input field.