Add structured logging interface
Initially this just provides a way to create category-specific
loggers and log at different warning levels using an API intended
to be compatible with Bunyan. The initial implementation just forwards
to console.log. The intention in future is to support logging to
other transports, etc. JSON logs in IndexedDB for troubleshooting
and opt-in telemetry.