Use only the metadata from .1password files
Fix a variety of possible sync issues due to the contents.js and
.1password files in a vault being out of sync by using only the
.1password files themselves for data about items currently in the vault.

This fixes issues including:

 * An error if a non-tombstone item was listed in contents.js but
   no such item existed in the vault.
 * Incorrect handling of items which had a newer updated-at timestamp
   in contents.js than the actual content in the .1password file.

The contents.js file is only used to get a list of tombstones
for deleted entries.

The contents.js file is now written by agile_keychain.Vault but only
for compatibility with the official 1Password apps.