Remove .1password files when item is deleted from vault
The official .1password apps delete the .1password file for an item
and leave behind only a tombstone entry in contents.js when an item
is removed.

Previously agile_keychain.Vault just erased the items fields in
the contents.js index file and the .1password file.

This makes the behavior consistent with the official apps by removing
the .1password file as well.

If a file is listed as a tombstone in the contents.js file but a
.1password file already exists then the deletion takes precedence.