Refactor logic to determine whether to sync changes for each item
The previous logic used item revisions to determine whether
to add an item to the sync queue and then later used item
last-updated timestamps to determine whether the item
had changed since the last sync in the local or cloud stores.

This refactors the logic to determine whether an item has
been added/updated/removed since the last sync in only
one place based upon a comparison of the last-synced revision
and the current revision.