Remove VFS.login(), logout(), isLoggedIn() methods
Authentication is now handled externally from the VFS implementations
in order to enable more control over the auth flow UI and consistent
handling of the auth flow across different VFS implementations.

This means that the VFS implementations just expose a URL where
the user can authenticate, if necessary and a method to set
the credentials used for API calls.