Fix issue where action buttons disappeared on click on Firefox under Mac
On Firefox and Safari for Mac, <button> elements do not receive focus
on click by default. When clicking on an action button, the focus
therefore went to the item field list component, resulting in
the item field thinking it had lost focus and hiding the action
buttons.

This fixes the issue by only hiding the action buttons when a
different field is selected, not when focus moves from a field
to one of the parent components of the field.

Fixes #47