Fix an issue where submit button labels were replaced with the username
On forms where there was a password field but no obvious username field,
the autofill logic picks the first text field and populates it with
the username. Fields with type 'submit' were incorrectly marked
as text fields and consequently their value (ie. label) was
replaced.

This affected the Google Account sign-in form amongst others.