Implement a generic OAuth flow independent of Dropbox
To make switching between multiple stores from different
Dropbox accounts work properly, refactor the OAuth flow
to use a generic implementation of the implicit flow
implemented in Passcards itself, rather than using
DropboxJS's default auth driver.

This provides an OAuth flow which should be usable
with any OAuth provider.

The access token received once OAuth is complete is
saved with the account's settings. This enables
the appropriate token to be used to sync with
a given account.

In future for added security
the token should be stored encrypted with the user's
data so that it is only accessible after login -
the flip-side of that is that after a password change
the old password will need to be used for the first login
on a given device, _or_ the user will need to be given
an opportunity to re-authenticate again.