Preserve app state better in Chrome when popup is dismissed
Preserve the state of the app by moving the root component's
DOM node to the background page when the popup is unloaded
and then move it into the new popup view when the popup
is next opened.

Previously the root component was unmounted when the popup
view was loaded and then re-rendered without preserving
any of the component's internal state tree on the next render.
Any app state that needed to persist had to be stored outside
the component tree.

This covered the most important app state but not many of the
smaller details such as the content of input fields or
which page of the setup view was active.

It would probably be possible to externalize all state
and there would be advantages to doing so, but this approach is
a simple and effective alternative.