A a TransitionContainer component to simplify custom entry/exit transitions
React's CSSTransitionGroup component listens for entry/exit lifecycle
hooks and notifies children via changes to the className property.

TransitionContainer is a similar concept except that it notifies
children of the transition state via explicit props and passes
callbacks for the children to notify the CSSTransitionGroup
when the animation completes, rather than waiting for
transitionend/animationend events.

Use this for animating the entry/exit of the dialog control.