Fix animations in Chrome extension
The requestAnimationFrame() wrapper designed to find a visible
view to schedule the rAF() call in to avoid Chrome's throttling
of rAF() for hidden pages (such as the background page where
the actual app rendering and logic lives) was invoking the call
in the background page - because window.document.hidden returns
false for that page.

Fix this by looking for a view which is (a) not hidden and (b)
not the background page.