Add transition timeouts for UnlockView slide animation
As noted in the React v0.14 release notes,
there are situations where CSS animation enter/leave events
are not emitted in various situations, one of which was when
the unlock view close animation happened whilst the UI was
hosted in the background page, where it lives when the
popup window is not visible. As a result, the slide-shut
animation never registered as having completed and
subsequent slide-open animations broke.

Updating to React v0.14 and specifying transition timeouts
resolves the issue.

Fixes #62