Remove 'webworker' from set of lib imports
The 'webworker' definitions lib contains declarations which conflict
with the 'dom' definitions lib.

Unless there is a way to tell TypeScript that a particular file or
function is used in a certain environment, it is easier to provide just
the missing definitions for scripts that are used in a Web Worker.