Work around CFX setting incorrect file attributes on locales.json
The CFX tool sets the file mode but not the file type on the external
attributes for the locales.json file in the .xpi archive. This caused
errors when the `jpm` tool processed the archive.

CFX should be eliminated entirely in future in favor of WebExtensions,
but until then, this works around the issue.