Use chrome.tabs.* API to open a window/tab for Dropbox OAuth authentication
Firefox disallows HTTP(S) URLs redirecting to moz-extension:// so it is
not possible for the Dropbox OAuth flow to redirect back to a page
served by the extension once OAuth completes.

Instead, redirect back to a dummy localhost URL, which ensures that we
don't send the access token to an external domain, and intercept this
using the chrome.tabs.* API.

The downside of this is that we now have to ask for the "tabs"
permission in the Chrome/WebExtension manifest.