Fix injection of content script into page in WebExtension
In Firefox v52a1, passing a 'data/dist/scripts/page_bundle.js' as the
URL to chrome.tabs.executeScript() resulted in the URL being resolved to
'data/data/dist/scripts/page_bundle.js' - ie.  apparently relative to
background.html. In Chrome on the other hand, it looks like the URL
is resolved relative to the root of the bundle.

Change the URL to an absolute path so that it resolves relative
to the root of the extension in both Chrome and Firefox