Do not attempt to re-use IndexedDB transactions
In Firefox v52, LocalStore#saveKeys() triggered a
TransactionInactiveError when trying to concurrently save the encryption
key and password hint to the database. The request to save the key
started a new IDB transaction and the request to save the hint attempted
to re-use that transaction.

This worked before when using Q.Promise but broke with the recent switch
to native Promises, possible due to the issues discussed in
https://github.com/dfahlander/Dexie.js/issues/317

Since transaction re-use is just an optimization, the problem is worked
around for the moment by just avoiding re-use.