Fix sporadic state mismatch error during login
The base64 state string required URL encoding and the Dropbox SDK did
not do this when constructing the auth URL.

Fix the problem by generating a random hex rather than base64 string
which is URL-safe.

Also fix an issue where the auth test was not actually run when running
UI tests.