Apply owner document hack when moving root view to background page
When the root view is 'saved' to the extension's background page, the
`_ownerDocument` hack needs to be applied as otherwise `_ownerDocument`
will refer to a 'dead object' (the closed popup window) and errors will
occur if React tries to access it.

This fixes the setup screen not advancing to the password vault list
after the user completes authentication in Firefox.